refactor: store ThemeData directly in cubit
